vedeo/requirements.txt
amqp==5.1.1
asgiref==3.7.2
backports.zoneinfo==0.2.1
billiard==4.1.0
bson==0.5.10
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.3
Django==4.2.4
django-cors-headers==4.2.0
django-crontab==0.7.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
dnspython==2.4.1
ffmpeg-python==0.2.0
future==0.18.3
idna==3.4
kombu==5.3.1
prompt-toolkit==3.0.39
psutil==5.9.5
pycparser==2.21
pymongo==4.4.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uWSGI==2.0.22
vine==5.0.0
wcwidth==0.2.6
